requirements.txt
torch
numpy
pillow
scikit-learn